Feat: Added PlacaMae, MemoriaRam, Fonte, Cpu, Gpu Controllers, services, and DTOs.

+++ src/main/java/com/projetoPC/dev/models/MemoriaRam.java
@@ -36,6 +36,6 @@ public class MemoriaRam {
     private String tipo;
 
     @Column(nullable = false, precision = 10, scale = 2)
-    private BigDecimal consumo;
+    private Integer consumo;
 
 }
